refactor: 일부 코드 정비
