+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   coverageDirectory: '../coverage',
   preset: 'ts-jest',
-  rootDir: 'src',
+  rootDir: './src',
   testEnvironment: 'node',
   transform: { '^.+\\.(ts|tsx)$': 'ts-jest' },
 };
