+++ src/http-client/http-client.type.ts
@@ -1,0 +1,4 @@
+import type { AxiosRequestConfig, AxiosResponse } from 'axios';
+
+export type HttpReqConfig = AxiosRequestConfig;
+export type HttpRes<T> = AxiosResponse<T>;
