+++ src/http-client/index.ts
@@ -1,3 +1,3 @@
 export { HttpClient } from './http-client';
 export { HttpState } from './http-state.enum';
-export type { HttpReqConfig, HttpRes } from './http-client.interface';
+export type { HttpReqConfig, HttpRes } from './http-client.type';
